--- aulaspi/README.md
@@ -1,2 +1,0 @@
-"# aulaspi" 
-Esse repositorio foi criado para conter os codigos fontes do projeto de eventos para aprender spring boot
