--- aulaspi/src/main/java/ifrn/pi/eventos/EventosApplication.java
@@ -1,13 +1,0 @@
-package ifrn.pi.eventos;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class EventosApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(EventosApplication.class, args);
-	}
-
-}
