--- aulaspi/src/main/java/ifrn/pi/eventos/controllers/indexController.java
@@ -1,16 +1,0 @@
-package ifrn.pi.eventos.controllers;
-
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class indexController {
-	
-	@RequestMapping("/")
-	public String index() {
-		System.out.println("Chamou o metudo");
-		return "reditect:/eventos";
-	}
-
-}
