--- aulaspi/src/main/java/ifrn/pi/eventos/repositories/EventoRepository.java
@@ -1,9 +1,0 @@
-package ifrn.pi.eventos.repositories;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ifrn.pi.eventos.models.Evento;
-
-
-public interface EventoRepository extends JpaRepository<Evento, Long>{
-
-}
