--- aulaspi/src/main/resources/application.properties
@@ -1,6 +1,0 @@
-spring.application.name=eventos
-spring.datasource.url=jdbc:mysql://localhost/eventos?serverTimezone=UTC
-spring.datasource.username=root
-spring.datasource.password=k.v15112004ds
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=true
