--- aulaspi/src/main/resources/templates/eventos/evento-adicionado.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-<meta charset="UTF-8">
-<!-- Compiled and minified CSS -->
-<link rel="stylesheet"
-	href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
-<title>Insert title here</title>
-</head>
-<body>
-	<div th:insert="~{menu :: menu}"></div>
-	<div class="container">
-		<h1>Evento adicionado com sucesso!</h1>
-	</div>
-</body>
-</html>
