--- aulaspi/src/main/resources/templates/eventos/formEvento.html
@@ -1,45 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-<meta charset="UTF-8">
-<!-- Compiled and minified CSS -->
-<link rel="stylesheet"
-	href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
-<title>Eventos</title>
-</head>
-<body>
-	<div th:insert="~{menu :: menu}"></div>
-
-	<div class="container">
-
-
-
-		<h1>Adicionar evento</h1>
-
-		<form action="/eventos" method="post">
-
-			<div>
-				<label>Nome:</label> <input type="text" name="nome">
-			</div>
-
-			<div>
-				<label>Local:</label> <input type="text" name="local">
-			</div>
-
-			<div>
-				<label>Data:</label> <input type="text" name="data">
-			</div>
-
-			<div>
-				<label>Horario:</label> <input type="text" name="horario">
-			</div>
-
-			<div>
-				<button class="waves-effect waves-light btn" type="submit">Salvar</button>
-				<button class="waves-effect waves-light btn" type="reset">Limpar</button>
-			</div>
-
-		</form>
-	</div>
-</body>
-</html>
