--- aulaspi/src/main/resources/templates/eventos/lista.html
@@ -1,41 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-<meta charset="UTF-8">
-<!-- Compiled and minified CSS -->
-<link rel="stylesheet"
-	href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
-<title>Eventos</title>
-</head>
-<body>
-	<div th:insert="~{menu :: menu}"></div>
-
-	<div class="container">
-
-
-
-		<h1>lista de eventos</h1>
-
-		<table>
-			<thead>
-				<tr>
-					<th>Id</th>
-					<th>Nome</th>
-					<th>Local</th>
-					<th>Data</th>
-					<th>Horário</th>
-				</tr>
-			</thead>
-			<tbody>
-				<tr th:each="e : ${eventos}">
-					<td th:text="${e.id}">1</td>
-					<td th:text="${e.nome}">joão aniversario</td>
-					<td th:text="${e.local}">um ai </td>
-					<td th:text="${e.data}">19/08</td>
-					<td th:text="${e.horario}">19:00</td>
-				</tr>
-			</tbody>
-		</table>
-	</div>
-</body>
-</html>
