--- aulaspi/src/main/resources/templates/home.html
@@ -1,18 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-<meta charset="UTF-8">
-<!-- Compiled and minified CSS -->
-<link rel="stylesheet"
-	href="https://cdnjs.cloudflare.com/ajax/libs/materialize/1.0.0/css/materialize.min.css">
-<title>Eventos</title>
-</head>
-<body>
-	<div th:insert="~{menu :: menu}"></div>
-	<div class="container">
-		<h1>Deu certo</h1>
-		<h1>Aplicação</h1>
-	</div>
-
-</body>
-</html>
