--- aulaspi/src/main/resources/templates/menu.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org">
-<head>
-<meta charset="UTF-8">
-
-<title>Eventos</title>
-</head>
-<body>
-	<nav th:fragment="menu" >
-		<div class="nav-wrapper container" >
-			<a href="#" class="brand-logo">Sistema de Eventos</a>
-			<ul id="nav-mobile" class="right hide-on-med-and-down">
-				<li><a href="/">Home</a></li>
-				<li><a href="eventos/form">Adicionar Evento</a></li>
-			</ul>
-		</div>
-	</nav>
-</body>
-</html>
