--- aulaspi/src/test/java/ifrn/pi/eventos/EventosApplicationTests.java
@@ -1,13 +1,0 @@
-package ifrn.pi.eventos;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class EventosApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
