+++ src/main/java/ifrn/pi/eventos/controllers/indexController.java
@@ -10,7 +10,7 @@ public class indexController {
 	@RequestMapping("/")
 	public String index() {
 		System.out.println("Chamou o metudo");
-		return "home";
+		return "redirect:/eventos";
 	}
 
 }
